0013 - [backend] add payment method retrieve api - Add Payment Retrieve API (GET for a specific userId) - Fix bug in ResourceMapper.convertPayment() extracting last 5 digits of card instead of 4 - Created com.bravo.payment package and moved Payment-specific classes there - Renamed PaymentDto to PaymentReadDto - service will need PaymentSaveDto soon to actually store encrypted CC number - Refactor: move classes shared between domains (user/payment) to new "common" package - Left 2 TODOs that need attention
