+++ src/main/java/com/bravo/payment/dao/repository/PaymentRepository.java
@@ -1,0 +1,16 @@
+package com.bravo.payment.dao.repository;
+
+import com.bravo.payment.dao.model.Payment;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
+import org.springframework.stereotype.Repository;
+
+import java.util.List;
+
+@Repository
+public interface PaymentRepository extends JpaRepository<Payment, String>, JpaSpecificationExecutor<Payment> {
+
+  Page<Payment> findByUserId(final String userId, Pageable pageRequest);
+}
