--- src/main/java/com/bravo/user/App.java
@@ -1,12 +1,0 @@
-package com.bravo.user;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class App {
-
-    public static void main(String[] args) {
-        SpringApplication.run(App.class, args);
-    }
-}
