--- src/main/java/com/bravo/user/model/dto/PaymentDto.java
@@ -1,14 +1,0 @@
-package com.bravo.user.model.dto;
-
-import java.time.LocalDateTime;
-import lombok.Data;
-
-@Data
-public class PaymentDto {
-
-  private String id;
-  private String cardNumberLast4;
-  private Integer expiryMonth;
-  private Integer expiryYear;
-  private LocalDateTime updated;
-}
