--- src/main/java/com/bravo/user/utility/DateUtil.java
@@ -1,12 +1,0 @@
-package com.bravo.user.utility;
-
-import java.time.format.DateTimeFormatter;
-import lombok.experimental.UtilityClass;
-
-@UtilityClass
-public class DateUtil {
-
-  public static final DateTimeFormatter DATE_FORMAT = DateTimeFormatter.ofPattern("yyyy-MM-dd");
-  public static final DateTimeFormatter DATE_TIME_FORMAT = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss");
-  public static final DateTimeFormatter TIME_FORMAT = DateTimeFormatter.ofPattern("HH:mm:ss");
-}
