+++ src/test/java/com/bravo/user/MapperArgConverter.java
@@ -1,8 +1,7 @@
 package com.bravo.user;
 
-import com.bravo.user.config.AppConfig;
+import com.bravo.config.AppConfig;
 import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.params.converter.ArgumentConversionException;
 import org.junit.jupiter.params.converter.SimpleArgumentConverter;
