+++ src/test/java/com/bravo/user/controller/UserControllerTest.java
@@ -1,11 +1,9 @@
 package com.bravo.user.controller;
 
-import com.bravo.user.App;
-import com.bravo.user.dao.model.User;
+import com.bravo.App;
 import com.bravo.user.model.dto.UserReadDto;
-import com.bravo.user.model.filter.UserFilter;
 import com.bravo.user.service.UserService;
-import com.bravo.user.utility.PageUtil;
+import com.bravo.common.utility.PageUtil;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
@@ -13,7 +11,6 @@ import javax.servlet.http.HttpServletResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.Mock;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
